make palette internal again
